move simple-button component to components directory
